task(GUIDEFRAME-15): selenium functionality moved to separate script
