+++ tutors_test.py
@@ -1,123 +1,27 @@
-from selenium import webdriver  # Importing the webdriver module from selenium and other modules
-from selenium.webdriver.common.by import By
-from selenium.webdriver.chrome.service import Service
-from selenium.webdriver.chrome.options import Options
-from selenium.webdriver.support.ui import WebDriverWait
-from selenium.webdriver.support import expected_conditions as EC
-from selenium.webdriver.common.action_chains import ActionChains
+from selenium_functions import * # Moved all selenium functions to external file
 import time  # Importing the time module for sleep functions
 from assembly import assemble  # Importing the assemble_clips function from assembly.py
 from audio import export_gtts, sleep_based_on_vo  # Importing the export_gtts and sleep_based_on_vo functions from audio.py
 from video import start_ffmpeg_recording, stop_ffmpeg_recording  # Importing the start_ffmpeg_recording and stop_ffmpeg_recording functions from video.py
 
 '''
-This file represents a more elaborate test script that uses the tutors website as a dynamic example. 
-This test will be expanded but for now should demonstrate the concept in a less static setting than the wiki test.
-Additionally, at the time of writing, this represents further refactoring where video and audio generation have been moved to separate files.
-This should further aid separation of concerns and hopefully render this more legible.
-Similarly, a potential improvement could be made in placing all selenium functions in a separate file but again, future work.
+As of GUIDEFRAME-15, selenium functions have moved to external script. This is to facilitate greater
+legibility of main file for ease of use. The intention is to approach the selenium material like an SDK-lite
+where documentation will outline the selenium functions and their args without the need to worry about try-catch,
+selenium syntax etc. This is simply the first step in that process.
 '''
 
-# Function to open a URL
-def open_url(driver, target):
-    driver.get(target)
-
-
-# Function to set window size
-def set_window_size(driver):
-    driver.maximize_window()
-
-
-# Function to find an element by its id
-def find_element(driver, id):
-    try:
-        element = WebDriverWait(driver, 10).until(
-            EC.presence_of_element_located((By.ID, id))
-        )
-        return element
-    except Exception as e:
-        print(f"Error finding element with ID '{id}': {e}")
-        return None
-
-# Function to scroll to an element using a href
-def scroll_to_element(driver, href):
-    try:
-        # Use WebDriverWait to ensure the element is present
-        element = WebDriverWait(driver, 10).until(
-            EC.presence_of_element_located((By.XPATH, f"//a[@href='{href}']"))
-        )
-        driver.execute_script("arguments[0].scrollIntoView()", element)
-
-
-    except Exception as e:
-        print(f"Error in scroll_to_element for href '{href}': {e}")
-
-
-# Function to hover over a href element and click it (ideal for link buttons)
-def hover_and_click(driver, href):
-    try:
-        # Use WebDriverWait to ensure the element is present
-        element = WebDriverWait(driver, 10).until(
-            EC.presence_of_element_located((By.XPATH, f"//a[@href='{href}']"))
-        )
-        # Use ActionChains to hover over the element
-        actions = ActionChains(driver)
-        actions.move_to_element(element).perform()
-
-        # Click the element
-        element.click()
-    except Exception as e:
-        print(f"Error in hover_and_click for href '{href}': {e}")
-
-
-# Function to hover over an href element (ideal for link buttons)
-def hover_over_element(driver, href):
-    try:
-        # Use WebDriverWait to ensure the element is present
-        element = WebDriverWait(driver, 10).until(
-            EC.presence_of_element_located((By.XPATH, f"//a[@href='{href}']"))
-        )
-        actions = ActionChains(driver)
-        actions.move_to_element(element).perform()
-
-    except Exception as e:
-        print(f"Error in hover_and_click for href '{href}': {e}")
-
-
-# Function to click an element using a CSS selector with error handling
-def click_element(driver, css_selector):
-    try:
-        element = WebDriverWait(driver, 10).until(
-            EC.presence_of_element_located((By.CSS_SELECTOR, css_selector))
-        )
-        element.click()
-    except Exception as e:
-        print(f"Error clicking element with selector '{css_selector}': {e}")
-
-
-# Function to run the modified Selenium test
-def run_selenium_test():
-    # Setting up with Chrome options and the ChromeDriver service
-    options = Options()
-    options.add_argument("usr/bin/google-chrome")
-
-    # Disable the "Chrome is being controlled by automated test software" banner
-    options.add_experimental_option("excludeSwitches", ["enable-automation"])
-    options.add_experimental_option("useAutomationExtension", False)
-
-    # Specify the path to ChromeDriver
-    service = Service('/opt/homebrew/bin/chromedriver')
-
-    # Initialize the WebDriver
-    driver = webdriver.Chrome(service=service, options=options)
+# Function to run the main walkthrough section
+def guideframe_script():
 
     # Main walkthrough logic here divided into steps etc (will hopefully be more legible after future refactors)
     #----------------------------------------------------------------------------------------------------------------------------------------------------------------------
 
     try:
         '''
-        Setup Open Tutors.dev and set window size etc
+        Setup - Setup driver and Open Tutors.dev and set window size etc
         '''
+        driver = driver_setup() # Initializing driver as the return value from the setup function in selenium script
         set_window_size(driver)
         open_url(driver, "https://tutors.dev")
         time.sleep(2)  # Give the page time to load
@@ -222,5 +126,5 @@ def run_selenium_test():
 
 # Main function to run the test and assemble the clips (now passing the number of steps to the assembly function)
 if __name__ == "__main__":
-    run_selenium_test()
+    guideframe_script()
     assemble(9)
